Add Home component to improve navigation
