+++ package.json
@@ -11,6 +11,8 @@
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router": "6.4.2",
+    "react-router-dom": "6.4.2",
     "react-scripts": "^5.0.1"
     
   },
