+++ src/components/Navbar.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Navbar = () => (
+  <div>Hello from Navbar</div>
+);
+
+export default Navbar;
