+++ src/components/TodoContainer.js
@@ -1,78 +1,20 @@
 /* eslint-disable no-param-reassign */
-import React, { useState, useEffect } from 'react';
-import { v4 as uuidv4 } from 'uuid';
-import Header from './Header';
-import InputTodo from './InputTodo';
-import TodoList from './TodoList';
+import React from 'react';
+import { Routes, Route } from 'react-router-dom';
+import NoMatch from '../pages/NoMatch';
+import About from '../pages/About';
+import Navbar from './Navbar';
+import Home from './home';
 
-function getInitialTodos() {
-  // getting stored items
-  const temp = localStorage.getItem('todos');
-  const savedTodos = JSON.parse(temp);
-  return savedTodos || [];
-}
+const TodoContainer = () => (
+  <>
+    <Routes>
+      <Route path="/" element={<Home />} />
+      <Route path="/about" element={<About />} />
+      <Route path="*" element={<NoMatch />} />
+    </Routes>
+  </>
 
-const TodoContainer = () => {
-  const [todos, setTodos] = useState(getInitialTodos());
-
-  useEffect(() => {
-    // storing todos items
-    const temp = JSON.stringify(todos);
-    localStorage.setItem('todos', temp);
-  }, [todos]);
-
-  const handleChange = (id) => {
-    setTodos((prevState) => prevState.map((todo) => {
-      if (todo.id === id) {
-        return {
-          ...todo,
-          completed: !todo.completed,
-        };
-      }
-      return todo;
-    }));
-  };
-
-  const delTodo = (id) => {
-    setTodos([
-      ...todos.filter((todo) => todo.id !== id),
-    ]);
-  };
-
-  const addTodoItem = (title) => {
-    const newTodo = {
-      id: uuidv4(),
-      title,
-      completed: false,
-    };
-    setTodos([...todos, newTodo]);
-  };
-
-  const setUpdate = (updatedTitle, id) => {
-    setTodos(
-      todos.map((todo) => {
-        if (todo.id === id) {
-          todo.title = updatedTitle;
-        }
-        return todo;
-      }),
-    );
-  };
-
-  return (
-    <div className="container">
-      <div className="inner">
-        <Header />
-        <InputTodo addTodoProps={addTodoItem} />
-        <TodoList
-          todos={todos}
-          handleChangeProps={handleChange}
-          deleteTodoProps={delTodo}
-          setUpdate={setUpdate}
-        />
-      </div>
-    </div>
-  );
-};
+);
 
 export default TodoContainer;
