+++ src/index.js
@@ -1,5 +1,6 @@
 import React, { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
+import { BrowserRouter as Router } from 'react-router-dom';
 import reportWebVitals from './reportWebVitals';
 
 import TodoContainer from './components/TodoContainer';
@@ -10,7 +11,9 @@ const root = createRoot(rootElement);
 
 root.render(
   <StrictMode>
-    <TodoContainer />
+    <Router>
+      <TodoContainer />
+    </Router>
   </StrictMode>,
 );
 reportWebVitals();
