+++ src/pages/About.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const About = () => (
+  <div>About</div>
+);
+
+export default About;
