+++ src/pages/NoMatch.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const NoMatch = () => (
+  <div>NoMatch</div>
+);
+
+export default NoMatch;
